docs(misc): add plugin examples to generators/executors pages
